Committing the first version of the field reference transformation which adds support for field references. This change requires the VarTypeTable to have a String key. Also adding support for FieldRefVarExpr across Uniqueness transformation.

+++ src/main/gov/nasa/jpf/symbc/veritesting/ast/transformations/typepropagation/ExprTypeVisitor.java
@@ -89,7 +89,7 @@ public class ExprTypeVisitor extends ExprMapVisitor implements ExprVisitor<Expre
 
     private String getConstantType(Expression op1) {
         if (op1 instanceof IntConstant) return "int";
-        else if (op1 instanceof RealConstant) return "float";
+        else if (op1 instanceof RealConstant) return "double";
         throw new IllegalArgumentException("trying to getConstantType for non-constant op, op = " + op1);
     }
